renames permalink model module
