+++ packages/models-library/src/models_library/api_schemas_webserver/permalinks.py
